[svn r18871] Minor - Adding if file_exists condition to avoid warning message

+++ index.php
@@ -1,4 +1,4 @@
-<?php // $Id: index.php 18470 2009-02-12 15:33:51Z juliomontoya $
+<?php // $Id: index.php 18871 2009-03-09 15:26:59Z juliomontoya $
  
 /*
 ==============================================================================
@@ -27,7 +27,7 @@
 /**
 *	@package dokeos.main
 * 	@author Patrick Cool <patrick.cool@UGent.be>, Ghent University, Refactoring
-* 	@version $Id: index.php 18470 2009-02-12 15:33:51Z juliomontoya $
+* 	@version $Id: index.php 18871 2009-03-09 15:26:59Z juliomontoya $
 *   @todo check the different @todos in this page and really do them
 * 	@todo check if the news management works as expected
 */
@@ -472,7 +472,8 @@ function display_anonymous_right_menu() {
 	}
 		
 	// includes for any files to be displayed below anonymous right menu
-	if (!file_exists($home.'home_notice_'.$user_selected_language.'.html') && file_get_contents($home.'home_notice.html')!='') {
+	
+	if (!file_exists($home.'home_notice_'.$user_selected_language.'.html') && file_exists($home.'home_notice.html') && file_get_contents($home.'home_notice.html')!='') {
 		echo '<div class="note">';
 		if (file_exists($home.'home_notice.html'))			
 			include ($home.'home_notice.html');
@@ -480,7 +481,7 @@ function display_anonymous_right_menu() {
 			include ($home_old.'home_notice.html');
 		}
 		echo '</div>';
-	} elseif(file_exists($home.'home_notice_'.$user_selected_language.'.html') && file_get_contents($home.'home_notice_'.$user_selected_language.'.html')!='') {
+	} elseif(file_exists($home.'home_notice_'.$user_selected_language.'.html') && file_exists($home.'home_notice_'.$user_selected_language.'.html') && file_get_contents($home.'home_notice_'.$user_selected_language.'.html')!='') {
 		echo '<div class="note">';
 		include($home.'home_notice_'.$user_selected_language.'.html'); 
 		echo '</div>';
